Restructures project
This commit introduces a significant restructuring of the project to accommodate a new design and overall improvements.

Removes old layouts and pages, replacing them with a new, more streamlined structure including updated styles and components.
This change allows for better maintainability and scalability.

Adds favicons, a new font and a stream banner.
Also replaces the logo.